make ProgressImpl a singleton and add tests in ToolTest
